refactor(tests): split useLLMd.test.ts into 5 focused files
Splits the 2269-line monolith into:
- useLLMdServers: initialization, demo mode, server type detection, component type detection, server status detection
- useLLMdServers-gpu: GPU extraction, autoscaler detection, gateway/prometheus namespace, model name extraction, status computation, error handling
- useLLMdServers-multicluster: multi-cluster loading, deployment filtering, server fields, refetch/auto-refresh, cleanup
- useLLMdModels: all useLLMdModels tests (initialization, demo, InferencePools, errors, refetch, isFailed, progressive loading)
- useLLMdServers-deep: additional deep coverage for consecutive failures, VA autoscaler, namespace matching, component type, status memoization

Fixes #8624
